Distinguish abilites by being increasable
Knacks don't have a level and cant be increased, ergo, abilities are not increasable per default. Only those with a level are.
