+++ module/data/item/_module.mjs
@@ -56,6 +56,7 @@ export {
 
 export {default as AbilityTemplate} from "./templates/ability.mjs";
 export {default as ClassTemplate} from "./templates/class.mjs";
+export {default as IncreasableAbilityTemplate} from "./templates/increasable-ability.mjs";
 export {default as ItemDescriptionTemplate} from "./templates/item-description.mjs";
 export {default as KnackTemplate} from "./templates/knack-item.mjs";
 export {default as LpIncreaseTemplate} from "./templates/lp-increase.mjs";
