+++ module/data/item/templates/ability.mjs
@@ -2,14 +2,11 @@ import ClassTemplate from "./class.mjs";
 import TargetTemplate from "./targeting.mjs";
 import ActionTemplate from "./action.mjs";
 import ED4E from "../../../config.mjs";
-import LpIncreaseTemplate from "./lp-increase.mjs";
 import LearnableTemplate from "./learnable.mjs";
 import PromptFactory from "../../../applications/global/prompt-factory.mjs";
-import LpSpendingTransactionData from "../../advancement/lp-spending-transaction.mjs";
 import RollPrompt from "../../../applications/global/roll-prompt.mjs";
 import AttackRollOptions from "../../roll/attack.mjs";
 import AbilityRollOptions from "../../roll/ability.mjs";
-const isEmpty = foundry.utils.isEmpty;
 
 /**
  * Data model template with information on Ability items.
@@ -17,14 +14,11 @@ const isEmpty = foundry.utils.isEmpty;
  * @property {object} source Class Source
  * @property {string} source.class class
  * @property {string} source.tier talent tier
- * @property {number} level rank
  * @mixes LearnableTemplate
- * @mixes LpIncreaseTemplate
  * @mixes TargetTemplate
  */
 export default class AbilityTemplate extends ActionTemplate.mixin(
   LearnableTemplate,
-  LpIncreaseTemplate,
   TargetTemplate
 ) {
   /** @inheritDoc */
@@ -66,15 +60,6 @@ export default class AbilityTemplate extends ActionTemplate.mixin(
         required: false,
         label:    this.labelKey( "Ability.Source.class" ),
         hint:     this.hintKey( "Ability.Source.class" )
-      } ),
-      level: new fields.NumberField( {
-        required: true,
-        nullable: false,
-        min:      0,
-        initial:  0,
-        integer:  true,
-        label:    this.labelKey( "Ability.rank" ),
-        hint:     this.hintKey( "Ability.rank" )
       } ),
       rollTypeDetails: new fields.SchemaField( {
         ability:       new fields.SchemaField( {}, {} ),
@@ -144,7 +129,7 @@ export default class AbilityTemplate extends ActionTemplate.mixin(
       rollingActorUuid: this.parentActor.uuid,
       abilityUuid:      this.parent.uuid,
       step:             {
-        base:      this.level + ( this.parentActor?.system.attributes[this.attribute]?.step ?? 0 ),
+        base:      this.rankFinal,
         modifiers: {},
       },
       karma:           rollOptions.karma,
@@ -170,11 +155,11 @@ export default class AbilityTemplate extends ActionTemplate.mixin(
   }
 
   /**
-   * The final rank of the ability (attribute step + level).
+   * The final rank of the ability (e.g. attribute + rank).
    * @type {number}
    */
   get rankFinal() {
-    return ( this.parentActor?.system.attributes[this.attribute]?.step ?? 0 ) + this.level;
+    return ( this.parentActor?.system.attributes[this.attribute]?.step ?? 0 );
   }
 
   /** @inheritDoc */
@@ -182,64 +167,9 @@ export default class AbilityTemplate extends ActionTemplate.mixin(
     return true;
   }
 
-  async adjustLevel( amount ) {
-    const currentLevel = this.level;
-    const updatedItem = await this.parent.update( {
-      "system.level": currentLevel + amount,
-    } );
-
-    if ( isEmpty( updatedItem ) ) {
-      ui.notifications.warn(
-        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
-      );
-      return;
-    }
-
-    return updatedItem;
-  }
-
-  /**
-   * @inheritDoc
-   */
-  async increase() {
-    if ( !this.isActorEmbedded ) return;
-
-    const promptFactory = PromptFactory.fromDocument( this.parent );
-    const spendLp = await promptFactory.getPrompt( "lpIncrease" );
-
-    if ( !spendLp
-      || spendLp === "cancel"
-      || spendLp === "close" ) return;
-
-    const currentLevel = this.level;
-
-    const updatedItem = await this.parent.update( {
-      "system.level": currentLevel + 1,
-    } );
-
-    if ( foundry.utils.isEmpty( updatedItem ) ) {
-      ui.notifications.warn(
-        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
-      );
-      return;
-    }
-
-    const updatedActor = await this.parent.actor.addLpTransaction(
-      "spendings",
-      LpSpendingTransactionData.dataFromLevelItem(
-        this.parent,
-        spendLp === "spendLp" ? this.requiredLpForIncrease : 0,
-        this.lpSpendingDescription,
-      ),
-    );
-
-    if ( foundry.utils.isEmpty( updatedActor ) )
-      ui.notifications.warn(
-        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
-      );
-
-    return this.parent;
-  }
+  /* -------------------------------------------- */
+  /*  Legend                                      */
+  /* -------------------------------------------- */
 
   async chooseTier( ) {
     const promptFactory = PromptFactory.fromDocument( this.parent );
@@ -273,7 +203,6 @@ export default class AbilityTemplate extends ActionTemplate.mixin(
       item.toObject(),
       foundry.utils.expandObject( createData ),
     );
-    if ( !createData?.system?.level ) itemData.system.level = 0;
     return ( await actor.createEmbeddedDocuments( "Item", [ itemData ] ) )?.[0];
   }
