+++ module/data/item/templates/increasable-ability.mjs
@@ -1,0 +1,139 @@
+import LpIncreaseTemplate from "./lp-increase.mjs";
+import PromptFactory from "../../../applications/global/prompt-factory.mjs";
+import LpSpendingTransactionData from "../../advancement/lp-spending-transaction.mjs";
+import AbilityTemplate from "./ability.mjs";
+const isEmpty = foundry.utils.isEmpty;
+
+/**
+ * Data model template with information on abilities that have rank and therefore can be increased with LP.
+ * @property {number} level rank
+ * @mixes LpIncreaseTemplate
+ */
+export default class IncreasableAbilityTemplate extends AbilityTemplate.mixin(
+  LpIncreaseTemplate,
+) {
+  /** @inheritDoc */
+  static defineSchema() {
+    const fields = foundry.data.fields;
+    return this.mergeSchema( super.defineSchema(), {
+      level: new fields.NumberField( {
+        required: true,
+        nullable: false,
+        min:      0,
+        initial:  0,
+        integer:  true,
+        label:    this.labelKey( "Ability.rank" ),
+        hint:     this.hintKey( "Ability.rank" )
+      } ),
+    } );
+  }
+
+  /* -------------------------------------------- */
+  /*  Getters                   */
+  /* -------------------------------------------- */
+
+  get baseRollOptions() {
+    const rollOptions = super.baseRollOptions;
+    rollOptions.updateSource( {
+      step:             {
+        base:      this.rankFinal,
+      },
+    } );
+
+    return rollOptions;
+  }
+
+  /** @inheritDoc */
+  get rankFinal() {
+    return super.rankFinal + this.level;
+  }
+
+  /* -------------------------------------------- */
+  /*  Legend                                      */
+  /* -------------------------------------------- */
+
+  async adjustLevel( amount ) {
+    const currentLevel = this.level;
+    const updatedItem = await this.parent.update( {
+      "system.level": currentLevel + amount,
+    } );
+
+    if ( isEmpty( updatedItem ) ) {
+      ui.notifications.warn(
+        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
+      );
+      return;
+    }
+
+    return updatedItem;
+  }
+
+  /**
+   * @inheritDoc
+   */
+  async increase() {
+    if ( !this.isActorEmbedded ) return;
+
+    const promptFactory = PromptFactory.fromDocument( this.parent );
+    const spendLp = await promptFactory.getPrompt( "lpIncrease" );
+
+    if ( !spendLp
+      || spendLp === "cancel"
+      || spendLp === "close" ) return;
+
+    const currentLevel = this.level;
+
+    const updatedItem = await this.parent.update( {
+      "system.level": currentLevel + 1,
+    } );
+
+    if ( foundry.utils.isEmpty( updatedItem ) ) {
+      ui.notifications.warn(
+        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
+      );
+      return;
+    }
+
+    const updatedActor = await this.parent.actor.addLpTransaction(
+      "spendings",
+      LpSpendingTransactionData.dataFromLevelItem(
+        this.parent,
+        spendLp === "spendLp" ? this.requiredLpForIncrease : 0,
+        this.lpSpendingDescription,
+      ),
+    );
+
+    if ( foundry.utils.isEmpty( updatedActor ) )
+      ui.notifications.warn(
+        game.i18n.localize( "ED.Notifications.Warn.Legend.abilityIncreaseProblems" )
+      );
+
+    return this.parent;
+  }
+
+  /** @inheritDoc */
+  static async learn( actor, item, createData ) {
+    if ( !item.system.canBeLearned ) {
+      ui.notifications.warn(
+        game.i18n.format( "ED.Notifications.Warn.Legend.cannotLearn", {itemType: item.type} )
+      );
+      return;
+    }
+    const itemData = foundry.utils.mergeObject(
+      item.toObject(),
+      foundry.utils.expandObject( createData ),
+    );
+    if ( !createData?.system?.level ) itemData.system.level = 0;
+    return ( await actor.createEmbeddedDocuments( "Item", [ itemData ] ) )?.[0];
+  }
+
+  /* -------------------------------------------- */
+  /*  Migrations                                  */
+  /* -------------------------------------------- */
+
+  /** @inheritDoc */
+  static migrateData( source ) {
+    super.migrateData( source );
+    // specific migration functions
+  }
+}
